Improved window, need to find a way to load images more easily.

+++ src/carlorolf/obstacles/BrickWall.java
@@ -9,7 +9,7 @@ import carlorolf.Images;
  * The brickwall sets the boundaries of the arena and keeps the arenaObjects inside.
  */
 public class BrickWall extends Wall {
-    public BrickWall(final double x, final double y, final int width, final int height,
+    public BrickWall(final int x, final int y, final int width, final int height,
                      final CollisionHandler collisionHandler, final Arena arena) {
         super(x, y, width, height, Images.getImage("bluebrick.png"), collisionHandler, arena);
     }
